requirements.txt
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.8.0
scipy==1.16.3
xgboost==3.1.2
lightgbm==4.6.0
optuna==4.3.0
torch==2.9.1
pytorch-tabnet==4.1.0
matplotlib==3.10.8
seaborn==0.13.2
plotly>=5.10.0,<6
sqlalchemy==2.0.45
pyyaml==6.0.3
joblib==1.5.3
openai==2.14.0
streamlit==1.52.2
python-dotenv==1.2.1
tqdm==4.67.1
pytest==9.0.2
mlflow==2.11.3
evidently==0.7.20
setuptools<81
# dev
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0
bandit==1.8.0
safety==3.3.1
pytest-cov==6.0.0
pytest-xdist==3.6.1
coverage==7.6.9